Increased test coverage to 85% + refactoring
There is a lot in this commit but this is because I added a large number of tests (about 80?)
Besides these tests I was prompted through writing them to add some new properties to classes
such as x_physical and y_physical to mesh. I additionally refactored a significant amount
including moving methods around and constants and functions. Despite these large changes very
little in the code was actually altered. The majority of new code was in the new tests.