upgraded to new shapely & fixed examples
shapely 2.0 has a huge breaking change, which is that Polygon objects
are now immutable and can not be subclassed. As such, I have had to make
polygon a property on the PySALEObject class. This is arguably a better
solution overall and should make shapely faster (so they say).

This commit also updates the examples and requirements.