sped up mesh creation & updated object repr
projecting polygons onto mesh was torturously slow. Now it is a lot more
efficient and quicker. All that happens is polygons are ordered by depth
and then filled in one by one selecting a region around the polygon and
only testing THAT instead of going through the entire mesh every time.
Obviously this won't make much of a difference for big polygons but when
there's lots of grains, times are cut drastically.

Also made the string rep of PySALEObjects more sensible.